+++ R/contributions_model_report.R
@@ -13,7 +13,7 @@ contributions_model_report <- report(class=c("contributions_model_report"))
 read.contributions_model_report <- function(report, ...) {
 
   . <- prob.TRUE <- truth <- score <- event <- timestamp <- expr_dt <- memb_amt <-
-    group_customer_no <- tck_amt <- customer_no <- order_dt <- NULL
+    group_customer_no <- tck_amt <- customer_no <- order_dt <- explanation <- NULL
 
   report$predictions <- read_cache("predictions","contributions_model") %>%
     collect %>% setDT %>%
@@ -59,14 +59,15 @@ process.contributions_model_report <- function(report, ...) {
 }
 
 #' @export
-#' @inheritParams send_xlsx
 #' @describeIn contributions_model_report Send spreadsheet for contributions_model_report
+#' @inheritParams send_email
 output.contributions_model_report <- function(report, subject = paste("Contributions model", Sys.Date()),
                                               body = paste("Sent by", Sys.info()["nodename"]), emails = config::get("tessiflow.email"), ...) {
 
-  send_xlsx(report$data, subject = subject, body = body, emails = emails, basename = "contributions_model",
+  send_xlsx(report$data, subject = subject, body = body, emails = emails, name = "contributions_model",
             currency = c("ticket_amt","census_median_income","iwave_capacity","iwave_donations","iwave_properties"))
   NextMethod()
 }
 
+#' @export
 write.contributions_model_report <- function(...) {NextMethod()}